Updated all helper functions
