+++ src/graphql/helpers/index.ts
@@ -1,3 +1,3 @@
-export * from './pagination'
-export * from './authorization'
+export * from './PaginationHelper'
+export * from './Authorization'
 export { PrismaHelper } from "./PrismaHelper"
