+++ src/graphql/types/server.types.ts
@@ -1,6 +1,6 @@
 import { ExpressContext } from 'apollo-server-express';
 import { AuthenticatedUser } from '../../apiBase/auth';
-import { IMe } from '../helpers/account';
+import { IMe } from '../helpers/Authorization';
 import { HouseholdLoaderType, PeopleFromHouseholdLoaderType } from '../loader';
 
 export type ReqContext = ExpressContext & {
